ADD: set props and emits for NodeManage
